dark_or_light_simulator.hex
Use your Calliope mini as a "weather station". Use the light sensor to measure brightness. When the sun is shining, the Calliope mini is happy and shows a smiley. If a cloud moves in front of the sun, it gets darker and the Calliope displays a sad smiley.
Tip: Depending on the ambient light, you have to adjust the numerical value for the light sensor.